+++ drift_dev/lib/src/analyzer/options.dart
@@ -95,9 +95,6 @@ class MoorOptions {
   @JsonKey(name: 'named_parameters_always_required', defaultValue: false)
   final bool namedParametersAlwaysRequired;
 
-  @JsonKey(name: 'new_sql_code_generation', defaultValue: false)
-  final bool newSqlCodeGeneration;
-
   @JsonKey(name: 'scoped_dart_components', defaultValue: false)
   final bool scopedDartComponents;
 
@@ -118,7 +115,6 @@ class MoorOptions {
     this.generateValuesInCopyWith = false,
     this.generateNamedParameters = false,
     this.namedParametersAlwaysRequired = false,
-    this.newSqlCodeGeneration = false,
     this.scopedDartComponents = false,
     this.modules = const [],
     this.sqliteAnalysisOptions,
@@ -141,7 +137,6 @@ class MoorOptions {
     required this.generateValuesInCopyWith,
     required this.generateNamedParameters,
     required this.namedParametersAlwaysRequired,
-    required this.newSqlCodeGeneration,
     required this.scopedDartComponents,
     required this.modules,
     required this.sqliteAnalysisOptions,