+++ drift_dev/lib/src/analyzer/options.g.dart
@@ -31,7 +31,6 @@ MoorOptions _$MoorOptionsFromJson(Map json) => $checkedCreate(
             'generate_values_in_copy_with',
             'named_parameters',
             'named_parameters_always_required',
-            'new_sql_code_generation',
             'scoped_dart_components'
           ],
         );
@@ -68,8 +67,6 @@ MoorOptions _$MoorOptionsFromJson(Map json) => $checkedCreate(
               $checkedConvert('named_parameters', (v) => v as bool? ?? false),
           namedParametersAlwaysRequired: $checkedConvert(
               'named_parameters_always_required', (v) => v as bool? ?? false),
-          newSqlCodeGeneration: $checkedConvert(
-              'new_sql_code_generation', (v) => v as bool? ?? false),
           scopedDartComponents: $checkedConvert(
               'scoped_dart_components', (v) => v as bool? ?? false),
           modules: $checkedConvert(
@@ -107,7 +104,6 @@ MoorOptions _$MoorOptionsFromJson(Map json) => $checkedCreate(
         'generateValuesInCopyWith': 'generate_values_in_copy_with',
         'generateNamedParameters': 'named_parameters',
         'namedParametersAlwaysRequired': 'named_parameters_always_required',
-        'newSqlCodeGeneration': 'new_sql_code_generation',
         'scopedDartComponents': 'scoped_dart_components',
         'modules': 'sqlite_modules',
         'sqliteAnalysisOptions': 'sqlite',