+++ drift_dev/lib/src/model/declarations/declaration.dart
@@ -41,12 +41,8 @@ abstract class MoorDeclaration extends Declaration {
 
 extension ToSql on MoorDeclaration {
   String exportSql(MoorOptions options) {
-    if (options.newSqlCodeGeneration) {
-      final writer = SqlWriter(options, escapeForDart: false);
-      return writer.writeSql(node);
-    } else {
-      return node.span!.text;
-    }
+    final writer = SqlWriter(options, escapeForDart: false);
+    return writer.writeSql(node);
   }
 }