+++ drift_dev/CHANGELOG.md
@@ -1,7 +1,8 @@
 ## 2.0.0-dev
 
-- Remove the `null_aware_type_converters` build option. Effectively, it is always
-  turned on now.
+- Removes the following build options, which are always turned on now:
+ - `null_aware_type_converters`
+ - `new_sql_code_generation`
 - Starting from this version, drift only supports generating non-nullable Dart code.
   You'll have to at least opt your database code into null-safety to use this and
   upcoming drift releases.